requirement.txt
pillow==9.4.0 
numpy==1.23.5
io
streamlit
random